server/requirements.txt
Flask
flask-cors
python-dotenv
firebase-admin
google-generativeai
gunicorn
